Delete confirmation && prevent empty search

+++ public/js/your-bugs.js
@@ -7,33 +7,38 @@ try {
 
     // Function call to delete a row from bug
     function deleteBug(tbl, curRow, bugId) {
-        let table = document.getElementById(tbl);
-        let rowCount = table.rows.length;
-        let req = new XMLHttpRequest();
-        let path = "/bug_tracker/all_bugs/deleteBug";
+        let del_confirm = confirm("Are you SURE you want to delete this?");
 
-        reqBody = JSON.stringify({bugId: bugId});
+        if (del_confirm) {
+            let table = document.getElementById(tbl);
+            let rowCount = table.rows.length;
+            let req = new XMLHttpRequest();
+            let path = "/bug_tracker/all_bugs/deleteBug";
 
-        req.open("POST", path, true);
-        req.setRequestHeader("Content-Type", "application/json");
-        req.addEventListener("load", () => {
-            if (req.status >= 200 && req.status < 400) {
-                for (let i = 0; i < rowCount; i++) {
-                    let row = table.rows[i];
-            
-                    if (row == curRow.parentNode.parentNode) {
-                        table.deleteRow(i);
+            reqBody = JSON.stringify({bugId: bugId});
+
+            req.open("POST", path, true);
+            req.setRequestHeader("Content-Type", "application/json");
+            req.addEventListener("load", () => {
+                if (req.status >= 200 && req.status < 400) {
+                    for (let i = 0; i < rowCount; i++) {
+                        let row = table.rows[i];
+                
+                        if (row == curRow.parentNode.parentNode) {
+                            table.deleteRow(i);
+                        }
                     }
+                    updateChartDelete();
+                } 
+                else {
+                    console.error("Delete request error");
                 }
-                updateChartDelete();
-            } 
-            else {
-                console.error("Delete request error");
-            }
-        });
+            });
 
-        req.send(reqBody);
+            req.send(reqBody);
+        }
     }
+
 
     /* VIEW ALL BUGS CLIENT SIDE ----------------------------------------------- */
 
@@ -42,7 +47,7 @@ try {
     viewAllButton.setAttribute('onclick', 'viewAllBugs()');
 
     function viewAllBugs() {
-        path = "/bug_tracker/all_bugs/viewAllBugs";
+        let path = "/bug_tracker/home/viewAll";
         let req = new XMLHttpRequest();
 
         req.open("POST", path, true);   
@@ -88,6 +93,12 @@ try {
 
     function searchBug() {
         let searchString = document.getElementById("search-input").value;
+
+        // End the search if the string is empty
+        if (searchString == '') {
+            return;
+        }
+
         let path = "/bug_tracker/all_bugs/searchBug";
         let req = new XMLHttpRequest();
         let reqBody = JSON.stringify({searchString: searchString});
@@ -105,7 +116,7 @@ try {
                 tableBody.innerHTML = '';
 
                 // if no results are found
-                if(bugsArray.length == 0) {
+                if (bugsArray.length == 0) {
                     let newRow = tableBody.insertRow(-1);
                     let summaryCell = document.createElement('td');
                     summaryCell.textContent = "No results found!";
